s4l and publication folders created
